Created ActivationK fitting function and removed previous one
Added header, source and tests for ActivationK function
Removed header, source and tests for Activation function
Updated CMakeLists

re #33090
