+++ Framework/CurveFitting/inc/MantidCurveFitting/Functions/ActivationK.h
@@ -15,13 +15,8 @@ namespace Mantid {
 namespace CurveFitting {
 namespace Functions {
 /**
-Provide Activation fit function interface to IFunction.
-I.e. if the unit is K, the function: y = = AttemptRate*exp(-Barrier/x)
-or if the unit is meV, the function: y = AttemptRate*exp(-e*Barrier/(1000 k_B x))
-Activation attributes:
-<UL>
-<LI> Unit - either K or meV (default K)</LI>
-</UL>
+Provide Activation fit function for data in Kelvin interface to IFunction.
+I.e. function: y = = AttemptRate*exp(-Barrier/x)
 
 Activation parameters:
 <UL>
@@ -30,21 +25,17 @@ Activation parameters:
 </UL>
 */
 
-class MANTID_CURVEFITTING_DLL Activation : public BackgroundFunction {
+class MANTID_CURVEFITTING_DLL ActivationK : public BackgroundFunction {
 public:
   /// overwrite IFunction base class methods
-  std::string name() const override { return "Activation"; }
+  std::string name() const override { return "ActivationK"; }
   void function1D(double *out, const double *xValues, const size_t nData) const override;
   void functionDeriv1D(API::Jacobian *out, const double *xValues, const size_t nData) override;
   const std::string category() const override { return "Muon\\MuonModelling"; }
-  void beforeFunctionSet() const;
 
 protected:
   /// overwrite IFunction base class method, which declare function parameters
   void init() override;
-
-private:
-  double getmeVConv() const;
 };
 
 } // namespace Functions
